Write leafentry code, and get things to compile.  Make it more compatible with -Wconversion.  Addresses #558.
git-svn-id: file:///svn/tokudb@3270 c7de825b-a66e-492c-adef-691d508d4ae1

+++ newbrt/mempool.h
@@ -25,7 +25,7 @@ struct mempool {
 
 /* initialize the memory pool with the base address and size of a
    contiguous chunk of memory */
-void toku_mempool_init(struct mempool *mp, void *base, int size);
+void toku_mempool_init(struct mempool *mp, void *base, size_t size);
 
 /* finalize the memory pool */
 void toku_mempool_fini(struct mempool *mp);
